Allows nodes to specify a custom OSC constructor
If a node's JSON definition contains the "makeConstructor" element, this must have expressions named "root" and "group" which contain code which creates a well-formed entry for the OSC packets array. The default would be:
  packets.add(OSC.GetCreateObjectAddr(),"ss", n.type, n.name);
and
  packets.add(OSC.GetCreateObjectAddr(),"sss", n.type, n.name, grpName);

Note that we have to assume that packets, n, grpName and so on continue to be defined as the GUI develops. An extended constructor would (for an object in the root) be something like:
   packets.add(OSC.GetCreateObjectAddr(),"ssif", n.type, n.name, node.memtype, node.length);
this adds an integer and a float to the OSC create object call